requirements.txt
# Core dependencies
langchain>=0.1.0
langchain-google-genai>=0.0.5
crewai>=0.14.0
python-dotenv>=1.0.0

# Remove conflicting pydantic version - let pip resolve automatically
# pydantic>=2.4.2  # This will be installed automatically by crewai

# Additional tools
requests>=2.31.0
